+++ flowmachine/flowmachine/features/subscriber/entropy.py
@@ -9,14 +9,17 @@ period.
 """
 
 from abc import ABCMeta, abstractmethod
-from typing import List
+from typing import List, Union
 
-from .metaclasses import SubscriberFeature
-from .contact_balance import ContactBalance
-from ..utilities.sets import EventsTablesUnion
-from ..utilities.subscriber_locations import SubscriberLocations
+
 from flowmachine.core import make_spatial_unit
 from flowmachine.core.spatial_unit import AnySpatialUnit
+from flowmachine.features.utilities.events_tables_union import EventsTablesUnion
+from flowmachine.features.utilities.subscriber_locations import SubscriberLocations
+from flowmachine.features.subscriber.contact_balance import ContactBalance
+from flowmachine.features.subscriber.metaclasses import SubscriberFeature
+from flowmachine.features.utilities.direction_enum import Direction
+from flowmachine.utils import make_where
 
 
 class BaseEntropy(SubscriberFeature, metaclass=ABCMeta):
@@ -88,8 +91,8 @@ class PeriodicEntropy(BaseEntropy):
         If provided, string or list of string which are msisdn or imeis to limit
         results to; or, a query or table which has a column with a name matching
         subscriber_identifier (typically, msisdn), to limit results to.
-    direction : {'in', 'out', 'both'}, default 'out'
-        Whether to consider calls made, received, or both. Defaults to 'out'.
+    direction : {'in', 'out', 'both'} or Direction, default Direction.BOTH
+        Whether to consider calls made, received, or both. Defaults to 'both'.
     hours : 2-tuple of floats, default 'all'
         Restrict the analysis to only a certain set
         of hours within each day.
@@ -120,7 +123,7 @@ class PeriodicEntropy(BaseEntropy):
         phase="hour",
         *,
         subscriber_identifier="msisdn",
-        direction="both",
+        direction: Union[str, Direction] = Direction.BOTH,
         hours="all",
         subscriber_subset=None,
         tables="all",
@@ -130,15 +133,14 @@ class PeriodicEntropy(BaseEntropy):
         self.start = start
         self.stop = stop
         self.subscriber_identifier = subscriber_identifier
-        self.direction = direction
+        self.direction = Direction(direction)
         self.hours = hours
 
-        if self.direction in {"both"}:
-            column_list = [self.subscriber_identifier, "datetime"]
-        elif self.direction in {"in", "out"}:
-            column_list = [self.subscriber_identifier, "datetime", "outgoing"]
-        else:
-            raise ValueError("{} is not a valid direction.".format(self.direction))
+        column_list = [
+            self.subscriber_identifier,
+            "datetime",
+            *self.direction.required_columns,
+        ]
 
         # extracted from the POSTGRES manual
         allowed_phases = (
@@ -183,16 +185,10 @@ class PeriodicEntropy(BaseEntropy):
     @property
     def _absolute_freq_query(self):
 
-        where_clause = ""
-        if self.direction != "both":
-            where_clause = (
-                f"WHERE outgoing = {'TRUE' if self.direction == 'out' else 'FALSE'}"
-            )
-
         return f"""
         SELECT subscriber, COUNT(*) AS absolute_freq FROM
         ({self.unioned_query.get_query()}) u
-        {where_clause}
+        {make_where(self.direction.get_filter_clause())}
         GROUP BY subscriber, EXTRACT( {self.phase} FROM datetime )
         HAVING COUNT(*) > 0
         """
@@ -322,8 +318,8 @@ class ContactEntropy(BaseEntropy):
         If provided, string or list of string which are msisdn or imeis to limit
         results to; or, a query or table which has a column with a name matching
         subscriber_identifier (typically, msisdn), to limit results to.
-    direction : {'in', 'out', 'both'}, default 'out'
-        Whether to consider calls made, received, or both. Defaults to 'out'.
+    direction : {'in', 'out', 'both'} or Direction, default Direction.BOTH
+        Whether to consider calls made, received, or both. Defaults to 'both'.
     hours : 2-tuple of floats, default 'all'
         Restrict the analysis to only a certain set
         of hours within each day.
@@ -355,7 +351,7 @@ class ContactEntropy(BaseEntropy):
         stop,
         *,
         subscriber_identifier="msisdn",
-        direction="both",
+        direction: Union[str, Direction] = Direction.BOTH,
         hours="all",
         subscriber_subset=None,
         tables="all",