+++ flowmachine/tests/test_dependency_graph.py
@@ -12,7 +12,9 @@ import IPython
 from io import StringIO
 
 from flowmachine.core import CustomQuery
+from flowmachine.core.context import get_db
 from flowmachine.core.dummy_query import DummyQuery
+from flowmachine.core.query_state import QueryStateMachine
 from flowmachine.core.subscriber_subsetter import make_subscriber_subsetter
 from flowmachine.features import daily_location, EventTableSubset
 
@@ -22,6 +24,10 @@ from flowmachine.core.dependency_graph import (
     unstored_dependencies_graph,
     plot_dependency_graph,
     store_queries_in_order,
+    dependencies_eligible_for_store,
+    queued_dependencies,
+    executing_dependencies,
+    query_progress,
 )
 
 
@@ -42,11 +48,11 @@ def test_print_dependency_tree():
     expected_output = textwrap.dedent(
         """\
         <Query of type: MostFrequentLocation, query_id: 'xxxxx'>
-          - <Query of type: PolygonSpatialUnit, query_id: 'xxxxx'>
+          - <Query of type: AdminSpatialUnit, query_id: 'xxxxx'>
              - <Table: 'geography.admin3', query_id: 'xxxxx'>
           - <Query of type: SubscriberLocations, query_id: 'xxxxx'>
              - <Query of type: JoinToLocation, query_id: 'xxxxx'>
-                - <Query of type: PolygonSpatialUnit, query_id: 'xxxxx'>
+                - <Query of type: AdminSpatialUnit, query_id: 'xxxxx'>
                    - <Table: 'geography.admin3', query_id: 'xxxxx'>
                 - <Query of type: EventsTablesUnion, query_id: 'xxxxx'>
                    - <Query of type: EventTableSubset, query_id: 'xxxxx'>
@@ -57,7 +63,7 @@ def test_print_dependency_tree():
                       - <Query of type: CustomQuery, query_id: 'xxxxx'>
                       - <Table: 'events.calls', query_id: 'xxxxx'>
                          - <Table: 'events.calls', query_id: 'xxxxx'>
-             - <Query of type: PolygonSpatialUnit, query_id: 'xxxxx'>
+             - <Query of type: AdminSpatialUnit, query_id: 'xxxxx'>
                 - <Table: 'geography.admin3', query_id: 'xxxxx'>
         """
     )
@@ -164,3 +170,86 @@ def test_store_queries_in_order():
     dummy5 = QueryWithStoreAssertions(dummy_param=["dummy5", dummy3, dummy4])
     graph = calculate_dependency_graph(dummy5)
     store_queries_in_order(graph)
+
+
+def test_dependencies_eligible_for_store():
+    """
+    Test that the set of only storeable dependencies is returned.
+    """
+
+    class UnStoreableQuery(DummyQuery):
+        @property
+        def table_name(self):
+            raise NotImplementedError("This dummy cannot be stored.")
+
+    dummy = DummyQuery(dummy_param="DUMMY")
+    unstoreable_dummy = UnStoreableQuery(dummy_param="UNSTOREABLE_DUMMY")
+
+    nested = DummyQuery(dummy_param=[dummy, unstoreable_dummy])
+    assert dependencies_eligible_for_store(nested) == {dummy, nested}
+
+
+def test_queued_dependencies(dummy_redis):
+    """
+    Test that only queued dependencies are returned.
+    """
+    dummy = DummyQuery(dummy_param="DUMMY")
+    queued_qsm = QueryStateMachine(dummy_redis, dummy.query_id, get_db().conn_id)
+    queued_qsm.enqueue()
+    stored_dummy = DummyQuery(dummy_param="STORED_DUMMY")
+    stored_dummy.store()
+    executing_dummy = DummyQuery(dummy_param="EXECUTING_DUMMY")
+    executing_qsm = QueryStateMachine(
+        dummy_redis, executing_dummy.query_id, get_db().conn_id
+    )
+    executing_qsm.enqueue()
+    executing_qsm.execute()
+
+    nested = DummyQuery(dummy_param=[dummy, stored_dummy, executing_dummy])
+    assert queued_dependencies(set([nested, dummy, stored_dummy, executing_dummy])) == [
+        dummy
+    ]
+
+
+def test_executing_dependencies(dummy_redis):
+    """
+    Test that only executing dependencies are returned.
+    """
+    dummy = DummyQuery(dummy_param="DUMMY")
+    queued_qsm = QueryStateMachine(dummy_redis, dummy.query_id, get_db().conn_id)
+    queued_qsm.enqueue()
+    stored_dummy = DummyQuery(dummy_param="STORED_DUMMY")
+    stored_dummy.store()
+    executing_dummy = DummyQuery(dummy_param="EXECUTING_DUMMY")
+    executing_qsm = QueryStateMachine(
+        dummy_redis, executing_dummy.query_id, get_db().conn_id
+    )
+    executing_qsm.enqueue()
+    executing_qsm.execute()
+
+    nested = DummyQuery(dummy_param=[dummy, stored_dummy, executing_dummy])
+    assert executing_dependencies(
+        set([nested, dummy, stored_dummy, executing_dummy])
+    ) == [executing_dummy]
+
+
+def test_query_progress(dummy_redis):
+    """
+    Test correct counts for dependency progress are returned.
+    """
+    dummy = DummyQuery(dummy_param="DUMMY")
+    queued_qsm = QueryStateMachine(dummy_redis, dummy.query_id, get_db().conn_id)
+    queued_qsm.enqueue()
+    stored_dummy = DummyQuery(dummy_param="STORED_DUMMY")
+    stored_dummy.store()
+    executing_dummy = DummyQuery(dummy_param="EXECUTING_DUMMY")
+    executing_qsm = QueryStateMachine(
+        dummy_redis, executing_dummy.query_id, get_db().conn_id
+    )
+    executing_qsm.enqueue()
+    executing_qsm.execute()
+
+    nested = DummyQuery(dummy_param=[dummy, stored_dummy, executing_dummy])
+    assert query_progress(nested) == dict(eligible=3, running=1, queued=1,)
+    nested.store()
+    assert query_progress(nested) == dict(eligible=0, running=0, queued=0,)